requirements.txt
# Core framework
agno

# LLM provider
anthropic

# DuckDuckGo search tool (used by agno.tools.duckduckgo)
duckduckgo-search>=6.3.2

# Article body extraction (used by agno.tools.newspaper4k)
newspaper4k
lxml
lxml_html_clean

# Environment variable loading
python-dotenv

# Data validation (Agno requires >=2.0)
pydantic>=2.0.0

# Testing
pytest
pytest-mock